requirements.txt

Django==2.1
mysqlclient==1.3.13
wheel==0.31.1
gunicorn==19.9.0
psycopg2==2.7.5
djangorestframework==3.7.7
